requirements.txt
certifi==2020.6.20
chardet==3.0.4
idna==2.10
nba-api==1.1.9
numpy==1.19.1
python-dateutil==2.8.1
requests==2.24.0
six==1.15.0
urllib3==1.25.10
